+++ cpp_driver.hpp
@@ -20,7 +20,7 @@ namespace nano_pow
 		void threshold_set (uint64_t threshold) override;
 		uint64_t threshold_get () const override;
 		void threads_set (unsigned threads) override;
-		unsigned threads_get () const override;
+		size_t threads_get () const override;
 		void memory_set (size_t memory) override;
 		uint64_t solve (std::array<uint64_t, 2> nonce) override;
 		void dump () const override;
@@ -28,7 +28,7 @@ namespace nano_pow
 		void barrier (std::unique_lock<std::mutex> & lock);
 		nano_pow::context context;
 		nano_pow::generator generator;
-		void run_loop (unsigned);
+		void run_loop (size_t);
 		std::atomic<unsigned> ready { 0 };
 		bool enable { false };
 		std::vector<std::unique_ptr<std::thread>> threads;
