+++ opencl_driver.hpp
@@ -37,7 +37,7 @@ namespace nano_pow
 		void threshold_set (uint64_t threshold) override;
 		uint64_t threshold_get () const override;
 		void threads_set (unsigned threads) override;
-		unsigned threads_get () const override;
+		size_t threads_get () const override;
 		void memory_set (size_t memory) override;
 		uint64_t solve (std::array<uint64_t, 2> nonce) override;
 		void dump () const override;
